Added Command and CommandTable classes
